requirements/local.txt
# Local development dependencies go here
-r base.txt

coverage==4.4.1
django-coverage-plugin==1.5.0

Sphinx==1.6.3
Werkzeug==0.12.2
django-test-plus==1.0.18
factory-boy==2.9.0

django-debug-toolbar==1.9.1

# improved REPL
ipdb==0.10.3

pytest-django==3.1.2
pytest-sugar==0.8.0
mock==2.0.0

pylint-common==0.2.5
pylint-celery==0.3
pylint-django==0.7.2
selenium==3.5.0

fabric==1.14.0
